+++ app/src/main/java/com/qrz/voicetriggerrecorder/record/RecordingCloseReason.kt
@@ -4,6 +4,6 @@ enum class RecordingCloseReason {
     EndSilence,
     ManualStop,
     ServiceStop,
-    Error,
-    ProcessDeath
+    ReadError,
+    Destroy
 }